Method of Characteristics - Waterway
Matlab Simulation scripts from Masters Thesis 2020

Method of Characteristics (MoC) with various friction models
The simulation is intended to study the behavior of pressure pulsations in the waterways of hydropower machinery
The MoC simulation simplifies the waterway to 1 dimension for practical purposes. 3D simulation is extremely costly, and was simply unfeasible in 2020.